feat: preload ProjectDetails imgs
So as to avoid them visibly loading as they transition in
the first time for a user who hasn't viewed the image resources yet.

They aren't mounted until they are transitioned in (or until
the element before them is transitioned in if using `Next Project`
button) so they weren't loading and being saved to cache until their
first transition.